+++ packages/preview-middleware-client/src/adp/quick-actions/add-new-subpage-quick-action-base.ts
@@ -11,6 +11,7 @@ import { DialogFactory, DialogNames } from '../dialog-factory';
 import { EnablementValidatorResult } from './enablement-validator';
 import { getTextBundle } from '../../i18n';
 import { SimpleQuickActionDefinitionBase } from './simple-quick-action-base';
+import { DIALOG_ENABLEMENT_VALIDATOR } from './dialog-enablement-validator';
 
 export const ADD_NEW_OBJECT_PAGE_ACTION = 'add-new-subpage';
 const CONTROL_TYPES = ['sap.f.DynamicPage', 'sap.uxap.ObjectPageLayout'];
@@ -60,7 +61,8 @@ export abstract class AddNewSubpageBase<ODataMetaModelType>
                     }
                     return undefined;
                 }
-            }
+            },
+            DIALOG_ENABLEMENT_VALIDATOR
         ]);
         this.existingPages = this.getApplicationPages();
     }